Revert "src/structures.lisp: fix fsbv for nested structures"
This commit was initially meant to address the problem with fsbv "and
foreign-typedefs with aggregate types" reported on cffi-devel on
2020-02-14.

However "fix translation of struct-typedefs" was commited on the
master branch on 2020-02-26 to address the problem instead.  This was
reverted on master on 2020-05-12 because of some issue with magicl
which can now be addressed.

After Reverting "src/structures.lisp: fix fsbv for nested structures"
the next few commits will reintroduce the reverted "fix tranlation of
struct-typedefs" commit and subsequently fix it.