implement DEFCALLBACK (NAME :EXPORT-P T) for suitable ECL/MKCL
This patch is targetted at patched versions of ECL/MKCL which support
the FFI syntax FFI:DEFCALLBACK (NAME CONVENTION EXPORT-P)

CFFI:DEFCALLBACK adds a new option, a boolean keyword :EXPORT-P to
access this feature of ECL/MKCL.  If the callback is compiled into a
shared library the C-NAME of NAME is exported and available as an
external symbol which can be called from other C code.

MKCL-1.1.11+ has refurbished the c-export-name declaration to export
names in DLLs via

 (declaim (mkcl:c-export-name (lisp-name c-name)))

So use that.

220107 - also support older patched mkcl 1.11.17+ for a while