src/libraries.lisp: (%do-load-foreign-library): work around canary fiasco
cffi canary stuff just doesnt work because load-foreign-library unloads
the library first everytime. And the symbol is never found.

when libfixposix (a shared library) goes about looking for lfp_buildinfo
as a part of this canary mechanism, it never finds the symbol because
the (shared - not static) library has just been unloaded.

[The failed call to dlsym also triggers a nasty ECL bufferoverflow]